# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tables:
    get:
      operationId: listTables
      tags: [tables]
      summary: List all tables
      responses:
        "200":
          description: List of tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Table"
    post:
      operationId: createTable
      tags: [tables]
      summary: Create a table
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
  /tables/{id}:
    get:
      operationId: getTable
      tags: [tables]
      summary: Get a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Table found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
        "404":
          description: Not found
    patch:
      operationId: updateTable
      tags: [tables]
      summary: Update a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Table"
    delete:
      operationId: deleteTable
      tags: [tables]
      summary: Delete a table
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
components:
  schemas:
...
    Table:
      type: object
      required: [id, number, capacity, status, createdAt]
      properties:
        id:
          type: integer
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]
        currentOrderId:
          type: ["integer", "null"]
        createdAt:
          type: string
    TableInput:
      type: object
      required: [number, capacity]
      properties:
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]
    TableUpdate:
      type: object
      properties:
        number:
          type: integer
        capacity:
          type: integer
        status:
          type: string
          enum: [available, occupied, reserved]

    # Customers
...